Detect/requirements.txt
# Core dependencies
pybullet>=3.2.5
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0

torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8

PyYAML>=6.0
pillow>=10.0.0
scipy>=1.10.0

tqdm>=4.65.0  # Progress bars
wandb>=0.15.0  # Experiment tracking 